+++ setup.py
@@ -82,7 +82,7 @@ setup(name='numina',
       package_data={'numina.tests.drps.1': ['drp.yaml'],
                    },
       ext_modules=[ext1, ext2, ext3],
-      entry_points={'console_scripts': ['numina = numina.user:main']},
+      entry_points={'console_scripts': ['numina = numina.user.cli:main']},
       setup_requires=['numpy'],
       install_requires=REQUIRES,
       zip_safe=False,
